login and  signup page with no functionality added .
This is modification for signup and login page in which I uploaded initial setup last time. I have used only html and css for style my signup and login form.
Login page is styled and it needs functionality added to it by using JS and on the other hand signup page needs more style